test: rename Add method at WorkerPool on Send, improve worker-pool tests
